ai-model/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
numpy==1.24.3
scikit-learn==1.3.0
pandas==2.0.3
pydantic==2.4.2
tensorflow-macos>=2.11.0; sys_platform == 'darwin' and platform_machine == 'arm64'
tensorflow>=2.11.0; (sys_platform != 'darwin' or platform_machine != 'arm64')
requests==2.31.0
pymongo==4.5.0
python-dotenv==1.0.0